+ LSP CMP lspSaga native_snippets | - luasnips
Added support for Test, but possibly needs work in addition to default
configs. Having issues with the default behavior resize windows using
<C-{arrow-keys}> and I've instead migrated to <C-->, <C-=>, <C-,>, and
<C-.>. Looking at the key commands it doesn't quite add up and appear
intuitive as the arrow keys were, but when looking at the keyboard (even
though we aren't using shift) from appearance it's the -, +, <, > keys
which do make since as - and + will decrease and increase height and the
greater than and less than symbols point in the directions in which they
move the window pane right and left.

Added a new directory for plugins we wish to disable from LazyVim, but
it's currently empty. Also created a new dir within plugins which is
plugins/lang. This will serve as a place to provide language specific
files with the name of the programming language or language server name
would be fitting for matching the intended naming convention. Similar
pattern being used in LazyVim. Overwrote the Golang defaults on the LSP
to also enable GC_Details which provides details on the Garbage
collector provided through the LSP. Attempting to use NativeSnippets and
away from Luasnips. Hoping this resolves an issue I've been having with
the tab key that seems to come from Luasnips. I did provide int he
comments a means to resolve this issue with Luasnips should we need to
go back.

I've made some keybind changes with cmp as well, and I'm looking to move
to having more explicit keys to have fewer issues with the lsp and cmp
engines assuming my intent. I've transitioned the complete command away
from <CR> and <S-CR> to <C-y> and <C-Y>. I've opted for using the
dracula.nvim plugin for a theme and away from using dracula-pro. Pro
required manually installing the plugin (not always available on all
machines) where dracula can be sourced from github. Seems to also be
more maintained and is meant for neovim rather than forcing the vim
plugin into Neovim. I also added a new keybind within Dap which resets
the DAP ui as I am often finding the panes are shifting and resizing,
and this command seems to resolve that issue. I've also added LSPSaga
and remapped LSP commands to it and have not encountered any keybind
overlaps as of yet.