Change to Go Please Read if issues are experienced in other go packages
In order to be able to build integration tests in a manner expected in
some private projects, we need to pass the build flags specified to the
LSP server. This is an "inclusive" rather than "exclusive" change and
shouldn't produce any issues, but something to check if an issue arises
in the future.